add get user function to search handler
