--- utils/types/type.ts
@@ -1,3 +1,0 @@
-export type Language = {
-  language: "English" | "Spanish" | "Mandarin";
-};
