+++ utils/types/user.ts
@@ -1,0 +1,12 @@
+export type Language = {
+  language: "English" | "Spanish" | "Mandarin";
+};
+
+export type UserInfo = {
+  phoneNumber: string;
+  firstName: string;
+  lastName: string;
+  picURL: string;
+  hash: number;
+  selectedLanguage: "English" | "Spanish" | "Mandarin";
+};
